refactor: move setBounds() and getBounds() types to the mixin
