+++ packages/dialog/src/vaadin-dialog-overlay-mixin.d.ts
@@ -7,6 +7,22 @@ import type { Constructor } from '@open-wc/dedupe-mixin';
 import type { OverlayMixinClass } from '@vaadin/overlay/src/vaadin-overlay-mixin.js';
 import type { DialogRenderer } from './vaadin-dialog.js';
 
+export type DialogOverlayBounds = {
+  top: number;
+  left: number;
+  width: number;
+  height: number;
+};
+
+export type DialogOverlayBoundsParam =
+  | DialogOverlayBounds
+  | {
+      top?: number | string;
+      left?: number | string;
+      width?: number | string;
+      height?: number | string;
+    };
+
 export declare function DialogOverlayMixin<T extends Constructor<HTMLElement>>(
   base: T,
 ): Constructor<DialogOverlayMixinClass> & Constructor<OverlayMixinClass> & T;
@@ -27,4 +43,14 @@ export declare class DialogOverlayMixinClass {
    * Custom function for rendering the dialog footer.
    */
   footerRenderer: DialogRenderer | null | undefined;
+
+  /**
+   * Retrieves the coordinates of the overlay.
+   */
+  getBounds(): DialogOverlayBounds;
+
+  /**
+   * Updates the coordinates of the overlay.
+   */
+  setBounds(bounds: DialogOverlayBoundsParam): void;
 }
